Put RekorClient and FulcioClient behind interfaces
This lets them be a little more extensible if the
underlying communication mechanism needs to change (grpc, http)
