+++ sigstore-java/src/main/java/dev/sigstore/rekor/client/RekorClient.java
@@ -15,66 +15,15 @@
  */
 package dev.sigstore.rekor.client;
 
-import static dev.sigstore.json.GsonSupplier.GSON;
-
-import com.google.api.client.http.ByteArrayContent;
-import com.google.api.client.http.GenericUrl;
-import com.google.api.client.http.HttpRequest;
-import com.google.api.client.http.HttpResponse;
-import com.google.api.client.http.HttpResponseException;
-import dev.sigstore.http.HttpClients;
-import dev.sigstore.http.HttpParams;
-import dev.sigstore.http.ImmutableHttpParams;
 import java.io.IOException;
 import java.net.URI;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
 import java.util.Optional;
 
 /** A client to communicate with a rekor service instance. */
-public class RekorClient {
-  public static final URI PUBLIC_GOOD_URI = URI.create("https://rekor.sigstore.dev");
-  public static final URI STAGING_URI = URI.create("https://rekor.sigstage.dev");
-
-  public static final String REKOR_ENTRIES_PATH = "/api/v1/log/entries";
-  public static final String REKOR_INDEX_SEARCH_PATH = "/api/v1/index/retrieve";
-
-  private final HttpParams httpParams;
-  private final URI uri;
-
-  public static RekorClient.Builder builder() {
-    return new RekorClient.Builder();
-  }
-
-  private RekorClient(HttpParams httpParams, URI uri) {
-    this.uri = uri;
-    this.httpParams = httpParams;
-  }
-
-  public static class Builder {
-    private HttpParams httpParams = ImmutableHttpParams.builder().build();
-    private URI uri = PUBLIC_GOOD_URI;
-
-    private Builder() {}
-
-    /** Configure the http properties, see {@link HttpParams}, {@link ImmutableHttpParams}. */
-    public Builder setHttpParams(HttpParams httpParams) {
-      this.httpParams = httpParams;
-      return this;
-    }
-
-    /** Base url of the remote rekor instance. */
-    public Builder setUri(URI uri) {
-      this.uri = uri;
-      return this;
-    }
-
-    public RekorClient build() {
-      return new RekorClient(httpParams, uri);
-    }
-  }
+public interface RekorClient {
+  URI PUBLIC_GOOD_URI = URI.create("https://rekor.sigstore.dev");
+  URI STAGING_URI = URI.create("https://rekor.sigstage.dev");
 
   /**
    * Put a new hashedrekord entry on the Rekor log.
@@ -82,54 +31,25 @@ public class RekorClient {
    * @param hashedRekordRequest the request to send to rekor
    * @return a {@link RekorResponse} with information about the log entry
    */
-  public RekorResponse putEntry(HashedRekordRequest hashedRekordRequest)
-      throws IOException, RekorParseException {
-    URI rekorPutEndpoint = uri.resolve(REKOR_ENTRIES_PATH);
+  RekorResponse putEntry(HashedRekordRequest hashedRekordRequest)
+      throws IOException, RekorParseException;
 
-    HttpRequest req =
-        HttpClients.newRequestFactory(httpParams)
-            .buildPostRequest(
-                new GenericUrl(rekorPutEndpoint),
-                ByteArrayContent.fromString(
-                    "application/json", hashedRekordRequest.toJsonPayload()));
-    req.getHeaders().set("Accept", "application/json");
-    req.getHeaders().set("Content-Type", "application/json");
+  /**
+   * Get an entry from the log
+   *
+   * @param hashedRekordRequest the entry to find
+   * @return the entry if found on the log, empty otherwise
+   */
+  Optional<RekorEntry> getEntry(HashedRekordRequest hashedRekordRequest)
+      throws IOException, RekorParseException;
 
-    HttpResponse resp = req.execute();
-    if (resp.getStatusCode() != 201) {
-      throw new IOException(
-          String.format(
-              Locale.ROOT,
-              "bad response from rekor @ '%s' : %s",
-              rekorPutEndpoint,
-              resp.parseAsString()));
-    }
-
-    URI rekorEntryUri = uri.resolve(resp.getHeaders().getLocation());
-    String entry = resp.parseAsString();
-    return RekorResponse.newRekorResponse(rekorEntryUri, entry);
-  }
-
-  public Optional<RekorEntry> getEntry(HashedRekordRequest hashedRekordRequest)
-      throws IOException, RekorParseException {
-    return getEntry(hashedRekordRequest.computeUUID());
-  }
-
-  public Optional<RekorEntry> getEntry(String UUID) throws IOException, RekorParseException {
-    URI getEntryURI = uri.resolve(REKOR_ENTRIES_PATH + "/" + UUID);
-    HttpRequest req =
-        HttpClients.newRequestFactory(httpParams).buildGetRequest(new GenericUrl(getEntryURI));
-    req.getHeaders().set("Accept", "application/json");
-    HttpResponse response;
-    try {
-      response = req.execute();
-    } catch (HttpResponseException e) {
-      if (e.getStatusCode() == 404) return Optional.empty();
-      throw e;
-    }
-    return Optional.of(
-        RekorResponse.newRekorResponse(getEntryURI, response.parseAsString()).getEntry());
-  }
+  /**
+   * Get an entry from the log
+   *
+   * @param UUID the uuid of the log entry
+   * @return the entry if found on the log, empty otherwise
+   */
+  Optional<RekorEntry> getEntry(String UUID) throws IOException, RekorParseException;
 
   /**
    * Returns a list of UUIDs for matching entries for the given search parameters.
@@ -139,31 +59,7 @@ public class RekorClient {
    * @param publicKeyFormat format of public key (one of 'pgp','x509','minisign', 'ssh', 'tuf')
    * @param publicKeyContent public key base64 encoded content
    */
-  public List<String> searchEntry(
+  List<String> searchEntry(
       String email, String hash, String publicKeyFormat, String publicKeyContent)
-      throws IOException {
-    URI rekorSearchEndpoint = uri.resolve(REKOR_INDEX_SEARCH_PATH);
-
-    HashMap<String, Object> publicKeyParams = null;
-    if (publicKeyContent != null) {
-      publicKeyParams = new HashMap<>();
-      publicKeyParams.put("format", publicKeyFormat);
-      publicKeyParams.put("content", publicKeyContent);
-    }
-    var data = new HashMap<String, Object>();
-    data.put("email", email);
-    data.put("hash", hash);
-    data.put("publicKey", publicKeyParams);
-
-    String contentString = GSON.get().toJson(data);
-    HttpRequest req =
-        HttpClients.newRequestFactory(httpParams)
-            .buildPostRequest(
-                new GenericUrl(rekorSearchEndpoint),
-                ByteArrayContent.fromString("application/json", contentString));
-    req.getHeaders().set("Accept", "application/json");
-    req.getHeaders().set("Content-Type", "application/json");
-    var response = req.execute();
-    return Arrays.asList(GSON.get().fromJson(response.parseAsString(), String[].class));
-  }
+      throws IOException;
 }
